image-resize/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-storage==1.13.0
google-cloud
wand